requirements.txt
mlx-whisper>=0.4.0
openai>=1.0.0
click>=8.0.0

# 話者分離用（--diarizeオプション使用時に必要）
pyannote.audio>=3.1
torch>=2.0.0
torchaudio>=2.0.0

